refactor(2736): campsite cleanups — DRY no-details 400 onto helper, shared poster, docs
Behavior-preserving campsite fixes from the extreme-campsite squad (separate
commit for traceability; all existing tests green, zero new suppressions).

- sync/batch.py: `_parse_error_response` no-details else-branch now calls the
  canonical `_record_all_events_failed(..., transient=True)` (field-for-field
  identical) instead of a hand-rolled loop — DRY on the queue-disposition helper
  and drops the function's cyclomatic complexity off the C901 ceiling.
- core/batch_partition.py: `create_aware_midpoint` docstring now states the
  edge-index (`0`) return for a 2-element same-key pair and the caller-must-clamp
  contract (the R1 invariant, now discoverable from the primitive).
- delivery/receivers.py: tighten the R1 clamp comment (only `0` reachable, not
  `len`); collapse a duplicated default literal in `_map_400` to `(body or {}).get`.
- tests/delivery: relocate the shared `_AllOrNothingBatchPoster`/`_FakeResponse`
  into `tests/delivery/_poison_batch_poster.py` (removes a cross-module private
  import); add `singleton_post_tuples()` to de-dup a comprehension.
- tests/core: hoist a `_stub_key` helper for the shape-blindness tests.

Out of scope, deferred to its own tracked issue (needs characterization tests
first): the batch_sync response-disposition triplication — Priivacy-ai/spec-kitty#2811.
